pyproject.toml
[tool.poetry]
name = "semgrep-search-db"
version = "1.3.0"
description = "Database creator for semgrep-search"
authors = ["Malte Heinzelmann <malte@cybaer.ninja>"]
readme = "README.md"
license = "GPL-3.0-or-later"
homepage = "https://github.com/hnzlmnn/semgrep-search-db"
packages = [{include = "sgsdb"}]

[tool.poetry.scripts]
sgs-db = "sgsdb.main:main"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
tinydb = "^4.8.0"
ruamel-yaml = "^0.17.40"
tqdm = "^4.66.2"
python-git-info = "^0.8.3"
tomli = "^2.0.1"
semgrep = "^1.70.0"
multiprocess = "^0.70.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]

include = ["**/*.py"]
target-version = "py310"

# Allow lines to be as long as 120 characters.
line-length = 120

# Group violations by containing file.
output-format = "full"

[tool.ruff.lint]

# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "F", "N", "W", "Q", "SLF", "SIM", "T20", "ERA", "COM",
    "YTT", "ANN", "S", "BLE", "FBT", "B", "COM", "C4", "DTZ", "T10", "FA",
    "ISC", "ICN", "INP", "PIE", "T20", "PYI",  "PT", "Q", "RSE", "RET",
    "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "FIX"]
ignore = ["BLE001", "ANN101", "FIX002", "ANN003", "N818"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Make sure autofix doesn't delete imports from __init__.py
ignore-init-module-imports = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "single"
